moved components out of UI folder
